# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for managing purchase transactions with multi-currency support",
        "title": "Purchase API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Check if the API is running and responding to requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Liveness probe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID for tracing",
                        "name": "X-Request-ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HealthResponse"
                        }
                    }
                }
            }
        },
        "/health/ready": {
            "get": {
                "description": "Check if the API is ready to handle requests. Verifies database connectivity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Readiness probe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID for tracing",
                        "name": "X-Request-ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HealthResponse"
                        }
                    },
                    "503": {
                        "description": "Service not ready - database unavailable",
                        "schema": {
                            "$ref": "#/definitions/api.HealthResponse"
                        }
                    }
                }
            }
        },
        "/purchases": {
            "post": {
                "description": "Create a new purchase transaction with description, date, and USD amount. Validates that description is 1-50 chars, date is in ISO 8601 format and not in the future, and amount is positive.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchases"
                ],
                "summary": "Create a new purchase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID for tracing",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "description": "Purchase details",
                        "name": "purchase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreatePurchaseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.PurchaseResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/purchases/{id}": {
            "get": {
                "description": "Retrieve a purchase by ID. Optionally provide a currency code to convert the amount using historical exchange rates from the transaction date or earlier (within 6 months).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchases"
                ],
                "summary": "Get a purchase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ISO 4217 currency code for conversion (e.g., EUR, GBP, JPY)",
                        "name": "currency",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Request ID for tracing",
                        "name": "X-Request-ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PurchaseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid currency, invalid UUID, or rate not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Purchase not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CreatePurchaseRequest": {
            "description": "Purchase creation details",
            "type": "object",
            "properties": {
                "amountUsd": {
                    "type": "string",
                    "example": "1500.00"
                },
                "description": {
                    "type": "string",
                    "example": "Office supplies purchase"
                },
                "transactionDate": {
                    "type": "string",
                    "example": "2026-06-15"
                }
            }
        },
        "api.ErrorResponse": {
            "description": "API error with code and message",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "Amount must be a positive number"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-06-18T14:30:00Z"
                }
            }
        },
        "api.HealthResponse": {
            "description": "Health status of the API",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "database not available"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "api.PurchaseResponse": {
            "description": "Complete purchase details with optional conversion",
            "type": "object",
            "properties": {
                "amountUsd": {
                    "type": "string",
                    "example": "1500.00"
                },
                "convertedAmount": {
                    "type": "string",
                    "example": "1275.00"
                },
                "convertedCurrency": {
                    "type": "string",
                    "example": "EUR"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-06-18T14:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Office supplies purchase"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "rate": {
                    "type": "string",
                    "example": "0.85"
                },
                "rateDate": {
                    "type": "string",
                    "example": "2026-06-15"
                },
                "transactionDate": {
                    "type": "string",
                    "example": "2026-06-15"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-06-18T14:30:00Z"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}